+++ petShop/petShop/Model/Receipt.cs
@@ -7,9 +7,9 @@ namespace petShop.Model
         public Guid Id { get; private set; }
         public User Seller { get; private set; }
         public DateTime DateTimeSale { get; private set; }
-        public int TotalAmount { get; private set; }
+        public decimal TotalAmount { get; private set; }
 
-        public Receipt(User seller, int totalAmount)
+        public Receipt(User seller, decimal totalAmount)
         {
             if (seller.Role != Role.Seller)
                 throw new InvalidOperationException("Only seller can issue fiscal receipt.");
